requirements.txt
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.16.0
async-timeout==4.0.3
bcrypt==4.0.1
beautifulsoup4==4.12.2
cachelib==0.12.0
cachetools==5.3.0
certifi==2024.2.2
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.3.2
click==8.1.7
configobj==5.0.8
configparser==5.3.0
cos-python-sdk-v5==1.9.23
crcmod==1.7
cryptography==40.0.2
Cython==0.29.34
decorator==5.1.1
dicttoxml==1.7.16
dnspython==2.3.0
docker==6.0.1
enum34==1.1.10
Flask==2.2.5
Flask-Session==0.4.0
flask-sock==0.7.0
Flask-SQLAlchemy==3.0.3
future==0.18.3
geoip2==4.8.0
gevent==24.2.1
gevent-websocket==0.10.1
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==3.0.3
h11==0.14.0
httplib2==0.22.0
idna==3.7
importlib-metadata==6.7.0
iniparse==0.5
ipaddress==1.0.23
IPy==1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==0.10.0
kitchen==1.2.6
MarkupSafe==2.1.5
mongo==0.2.0
natsort==8.4.0
oauthlib==3.2.2
oss2==2.17.0
packaging==23.1
paramiko==3.4.0
peewee==3.16.2
pillow==10.3.0
protobuf==4.22.3
psutil==5.9.5
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyasyncore==1.0.4
pycparser==2.21
pycryptodome==3.17
pycurl==7.45.2
Pygments==2.15.1
pyinotify==0.9.6
pymongo==4.6.3
PyMySQL==1.0.3
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
pypdf==3.17.0
PySocks==1.7.1
pytz==2023.3
pyudev==0.24.1
pyxattr==0.8.1
PyYAML==6.0.1
qiniu==7.10.0
qrcode==7.4.2
redis==4.5.4
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
simple-websocket==0.10.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.10
supervisor==4.2.5
typing-extensions==4.7.1
upyun==2.5.5
uritemplate==4.1.1
urlgrabber==4.1.0
urllib3==1.26.18
websocket-client==1.5.1
Werkzeug==2.2.3
wheel==0.38.1
wsproto==1.2.0
xmltodict==0.13.0
zipp==3.15.0
zope.event==5.0
zope.interface==6.2
pandas==2.2.2
Brotli==1.1.0
rarfile==4.2
pyotp==2.9.0
unrar==0.4
sphinxapi-py3==2.1.11
jmxquery==0.6.0
billiard==4.2.0
lxml==5.0.0
docxtpl==0.16.8
ntplib==0.4.0
pyasynchat==1.0.4
toml==0.10.2
boto3==1.34.102
distro==1.9.0
pymssql==2.3.0
distro==1.9.0
pymssql==2.3.0